Updated bar charts rendered through mapping
